+++ docs/en_US/changelog.md
@@ -1,2 +1,5 @@
 ###2 Nov 21 
+  * Ajout de l'action 'activer et lancer', avec les tags définis
+
+###2 Nov 21 
   * Beta release